* tag 'android-7.0.0_r1':(82 commits)
  Allow system_server access to NFC reader mode API.
  Allow system_server access to NFC reader mode API.
  Allow peripheral/WiFi handover in provision mode.
  Import translations. DO NOT MERGE
  Verify setForegroundDispatch caller is in foreground. (DO NOT MERGE)
  Verify setForegroundDispatch caller is in foreground.
  Verify setForegroundDispatch caller is in foreground.
  Import translations. DO NOT MERGE
  Import translations. DO NOT MERGE
  Fix API review comments, add remove callback.
  Only add service when object is complete initialized.
  Don't use UserHandle.CURRENT when fetching user restrictions.
  Nfc shouldn't initiate bluetooth connection if disallowed.
  Allow foreground override in provision mode.
  Don't disable NFC in airplane mode.
  Keyboard Shortcuts: remove unnecessary method
  Import translations. DO NOT MERGE
  Import translations. DO NOT MERGE
  Import translations. DO NOT MERGE
  Fix reference to renamed string.
  ...

+++ nxp/src/com/android/nfc/dhimpl/NativeNfcManager.java
@@ -32,6 +32,7 @@ import android.nfc.ErrorCodes;
 import android.nfc.tech.Ndef;
 import android.nfc.tech.TagTechnology;
 import android.util.Log;
+import com.android.nfc.NfcDiscoveryParameters;
 
 import java.io.File;
 
@@ -75,7 +76,7 @@ public class NativeNfcManager implements DeviceHost {
     public static final String INTERNAL_TARGET_DESELECTED_ACTION = "com.android.nfc.action.INTERNAL_TARGET_DESELECTED";
 
     /* Native structure */
-    private int mNative;
+    private long mNative;
 
     private final DeviceHostListener mListener;
     private final Context mContext;
